feat: add generic page tabs component
