+++ src/components/page-tabs/index.ts
@@ -1,0 +1,2 @@
+export * from "./page-tabs";
+export * from "./page-tabs.types";
